wp-now: Always run in the same directory ($WPNOW_DIR)
Before, wp-now would run from a temporary directory when a plugin ZIP
artefact was available in GitHub, whereas it would run from the current
directory if it had to resort to performing a fresh build.

Now we drop the distinction: when a plugin ZIP artefact is available,
unzip it into the current directory, even if that means overwriting
source files.

By default, wp-now will run from the current directory, but it will
honour the $WPNOW_DIR environment variable.